rename demo and rearrange
