+++ .gitignore
@@ -5,3 +5,5 @@ build
 .php_cs.cache
 tests/Fake/fake-app/var/log/*
 tests/Fake/fake-app/var/tmp/*
+example/var/tmp/*
+example/var/log/*
