--- docs/demo-app/bootstrap/web.php
@@ -1,8 +1,0 @@
-<?php
-/**
- * This file is part of the BEAR.Package package.
- *
- * @license http://opensource.org/licenses/MIT MIT
- */
-$context = PHP_SAPI === 'cli' ? 'cli-hal-app' : 'hal-app';
-require __DIR__ . '/bootstrap.php';
