--- docs/demo-app/var/www/.htaccess
@@ -1,14 +1,0 @@
-order deny,allow
-
-<IfModule mod_rewrite.c>
-    # turn on rewriting
-    RewriteEngine On
-    
-    # for all files not found in the file system,
-    # reroute to "index.php" bootstrap script,
-    # keeping the query string intact.
-    RewriteCond %{REQUEST_FILENAME} !-d
-    RewriteCond %{REQUEST_FILENAME} !-f
-    RewriteCond %{REQUEST_FILENAME} !favicon.ico$
-    RewriteRule ^(.*)$ /index.php [QSA,L]
-</IfModule>
