--- docs/demo-app/var/www/index.php
@@ -1,6 +1,0 @@
-<?php
-
-//require dirname(__DIR__) . '/lib/xhprof.php';
-
-$context = 'prod-hal-app';
-require dirname(dirname(__DIR__)) . '/bootstrap/bootstrap.php';
