+++ example/README.md
@@ -1,0 +1,31 @@
+# Demo
+
+This is a minimal BEAR.Sunday application with BEAR.Package.
+ 
+## Install and Test
+```
+composer install
+./vendor/bin/phpunit
+```
+
+## Run
+
+
+console access
+
+```
+php public/index get /
+```
+
+web access
+
+```
+php -S 127.0.0.1:8080 -t public
+```
+
+batch command
+
+```
+php bin/run.php
+```
+
