+++ example/composer.json
@@ -1,0 +1,17 @@
+{
+    "name": "my-vendor/hello-package",
+    "require": {
+        "bear/sunday": "^1.2"
+    },
+    "autoload": {
+        "psr-4": {
+            "MyVendor\\HelloWorld\\": "src"
+        }
+    },
+    "require-dev": {
+        "phpunit/phpunit": "^6.3"
+    },
+    "scripts": {
+        "serve": "php -S 127.0.0.1:8080 -t public"
+    }
+}
