+++ example/src/Module/AppModule.php
