+++ example/src/Resource/Page/Api/User.php
@@ -4,7 +4,7 @@
  *
  * @license http://opensource.org/licenses/MIT MIT
  */
-namespace MyVendor\MyApp\Resource\App;
+namespace MyVendor\MyApp\Resource\Page\Api;
 
 use BEAR\RepositoryModule\Annotation\Cacheable;
 use BEAR\Resource\Annotation\Embed;
@@ -17,14 +17,16 @@ use BEAR\Resource\ResourceObject;
 class User extends ResourceObject
 {
     /**
-     * @Link(rel="profile", href="/profile{?id}")
-     * @Embed(rel="website", src="/website{?id}")
-     * @Embed(rel="contact", src="/contact{?id}")
+     * @Link(rel="profile", href="/api/profile{?id}")
+     * @Embed(rel="website", src="/api/website{?id}")
+     * @Embed(rel="contact", src="/api/contact{?id}")
      */
     public function onGet($id)
     {
-        $this['id'] = $id;
-        $this['name'] = 'Akihito Koriyama';
+        $this->body += [
+            'id' => $id,
+            'name' => 'Koriym'
+        ];
 
         return $this;
     }
