+++ example/src/Resource/Page/Api/User/Friend.php
